Part 3 of PC9801: move video routines into pc9801.cpp file (nw)
